# lab02/user_service_openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "User Service API",
      "version": "0.1.0"
    },
    "paths": {
      "/auth/token": {
        "post": {
          "tags": [
            "auth"
          ],
          "summary": "Authenticate",
          "operationId": "login_for_access_token_auth_token_post",
          "requestBody": {
            "content": {
              "application/x-www-form-urlencoded": {
                "schema": {
                  "$ref": "#/components/schemas/Body_login_for_access_token_auth_token_post"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/AuthResponse"
                  }
                }
              }
            },
            "401": {
              "description": "Invalid creds"
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/users": {
        "get": {
          "tags": [
            "users"
          ],
          "summary": "Get Users",
          "operationId": "get_users_users_get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/UserResponse"
                    },
                    "type": "array",
                    "title": "Response Get Users Users Get"
                  }
                }
              }
            }
          },
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ]
        },
        "post": {
          "tags": [
            "users"
          ],
          "summary": "Create a new user",
          "operationId": "create_user_users_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCreate"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/UserResponse"
                  }
                }
              }
            },
            "201": {
              "description": "User successfully created"
            },
            "400": {
              "description": "User already exists"
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/users/search": {
        "post": {
          "tags": [
            "users"
          ],
          "summary": "Search User",
          "operationId": "search_user_users_search_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchUser"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/UserResponse"
                    },
                    "type": "array",
                    "title": "Response Search User Users Search Post"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          },
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ]
        }
      },
      "/users/{user_id}": {
        "get": {
          "tags": [
            "users"
          ],
          "summary": "Get User",
          "operationId": "get_user_users__user_id__get",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "user_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "User Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "User found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/UserResponse"
                  }
                }
              }
            },
            "401": {
              "description": "Invalid token"
            },
            "404": {
              "description": "User not found"
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "users"
          ],
          "summary": "Delete current user",
          "operationId": "delete_user_users__user_id__delete",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "user_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "User Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response"
            },
            "204": {
              "description": "User deleted"
            },
            "401": {
              "description": "No access"
            },
            "404": {
              "description": "User not found"
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "AuthResponse": {
          "properties": {
            "access_token": {
              "type": "string",
              "title": "Access Token"
            },
            "token_type": {
              "type": "string",
              "title": "Token Type"
            }
          },
          "type": "object",
          "required": [
            "access_token",
            "token_type"
          ],
          "title": "AuthResponse"
        },
        "Body_login_for_access_token_auth_token_post": {
          "properties": {
            "grant_type": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^password$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Grant Type"
            },
            "username": {
              "type": "string",
              "title": "Username"
            },
            "password": {
              "type": "string",
              "title": "Password"
            },
            "scope": {
              "type": "string",
              "title": "Scope",
              "default": ""
            },
            "client_id": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Client Id"
            },
            "client_secret": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Client Secret"
            }
          },
          "type": "object",
          "required": [
            "username",
            "password"
          ],
          "title": "Body_login_for_access_token_auth_token_post"
        },
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
        "SearchUser": {
          "properties": {
            "fields": {
              "items": {
                "type": "string"
              },
              "type": "array",
              "title": "Fields"
            },
            "value": {
              "type": "string",
              "title": "Value"
            }
          },
          "type": "object",
          "required": [
            "fields",
            "value"
          ],
          "title": "SearchUser"
        },
        "UserCreate": {
          "properties": {
            "login": {
              "type": "string",
              "title": "Login"
            },
            "password": {
              "type": "string",
              "title": "Password"
            },
            "name": {
              "type": "string",
              "title": "Name"
            },
            "surname": {
              "type": "string",
              "title": "Surname"
            },
            "email": {
              "type": "string",
              "title": "Email"
            },
            "age": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Age"
            }
          },
          "type": "object",
          "required": [
            "login",
            "password",
            "name",
            "surname",
            "email"
          ],
          "title": "UserCreate"
        },
        "UserResponse": {
          "properties": {
            "id": {
              "type": "integer",
              "title": "Id"
            },
            "login": {
              "type": "string",
              "title": "Login"
            },
            "name": {
              "type": "string",
              "title": "Name"
            },
            "surname": {
              "type": "string",
              "title": "Surname"
            },
            "email": {
              "type": "string",
              "title": "Email"
            },
            "age": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Age"
            }
          },
          "type": "object",
          "required": [
            "id",
            "login",
            "name",
            "surname",
            "email"
          ],
          "title": "UserResponse"
        },
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": [
            "loc",
            "msg",
            "type"
          ],
          "title": "ValidationError"
        }
      },
      "securitySchemes": {
        "OAuth2PasswordBearer": {
          "type": "oauth2",
          "flows": {
            "password": {
              "scopes": {
  
              },
              "tokenUrl": "token"
            }
          }
        }
      }
    }
  }